Split homeFactory to controller, resolver, type
