+++ controllers/home/types.ts
@@ -1,0 +1,9 @@
+export type HomePage = {
+  data: {
+    privatePosts: {
+      popularStudyPosts: StudyOverview[];
+      myStudyPosts?: StudyOverview[];
+    };
+    studyPosts: StudyOverview[];
+  };
+};
